pyproject.toml
[tool.poetry]
name = "myapp-textsuite"
version = "2.0.0"
description = "A web app which uses various api to achieve text extraction, text translation and text querying."
authors = ["kaushalmeena <kaushal.meena1996@gmail.com>"]
readme = "ReadMe.md"
repository = "https://github.com/kaushalmeena1996/myapp-textsuite"
packages = [
    { include = "textsuite" }
]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.2"
requests = "^2.23.0"
Flask_Compress = "^1.4.0"
user_agents = "^2.1"
googletrans = "^2.4.0"
gunicorn = "^20.0.4"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
flake8-black = "^0.1.1"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
flake8-bandit = "^2.1.2"

[tool.poetry.scripts]
start = "wsgi:start"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
